requirements.txt
grpcio_tools==1.49.1
loguru==0.6.0
matplotlib==3.6.0
numpy==1.23.3
phe==1.5.0
protobuf==4.22.1
pydotplus==2.0.2
PyNaCl==1.5.0
PyYAML==6.0
scikit_learn==1.2.2
tqdm==4.65.0
websockets==10.4